+++ src/tokens/molecules/recovery-codes.ts
@@ -1,0 +1,4 @@
+import recoveryCodes from '../../../tokens/molecule/recovery-codes.json';
+import { flattenTokens } from '../utils';
+
+export const recoveryCodesTokens = flattenTokens(recoveryCodes as never);